% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.loocv.R
\name{autoloocv.llr}
\alias{autoloocv.llr}
\title{Auto selection of bandwidth using metaheuristic algorithms}
\usage{
autoloocv.llr(
  x,
  y,
  weight,
  kernel = "epanechnikov",
  bw_range = "auto",
  approx = FALSE,
  epsilon = 0.05,
  control = list(numPopulation = "auto", maxIter = "auto", Vmax = 2, ci = 1.49445, cg =
    1.49445, w = 0.729),
  algorithm = "PSO",
  seed = 1
)
}
\arguments{
\item{x}{a numeric vector or matrix of x data}

\item{y}{a numeric vector of y data corresponding to \code{x} data}

\item{weight}{a numeric vector of \code{length(x)} for weight of each data point}

\item{kernel}{kernel type used; supported are 'epanechnikov', "rectangular", "triangular", "quartic", "triweight", "tricube", "cosine", "gauss", "logistic". "sigmoid" and "silverman".}

\item{bw_range}{range of bandwidth for the metaheuristic algorithm to search}

\item{approx}{boolean flag: if \code{true} kdtree approximation is used.}

\item{epsilon}{margin of error allowed for llr approximation using kdtree. Only used when \code{approx = TRUE}.}

\item{control}{control parameter from R package  metaheuristicOpt}

\item{algorithm}{algorithm parameter from R pacakage  metaheuristicOpt}

\item{seed}{seed to use for algorithm}
}
\description{
Auto selection of bandwidth using metaheuristic algorithms
}
\examples{
 
\dontrun{
n <- 1000
x <- runif(n, 0 , 1)
y <- sin(x) + rnorm(n, 0, 0.2)
w <- rep(1/n , n)
## Optimization using Grasshopper Optimisation Algorithm
algorithm <- 'GOA'
control <- list(numPopulation = 10, maxIter = 100)
h <- autoloocv.llr(x, y, w, control = control , algorithm = algorithm)
}
}
