% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Splitting 1D/2D data into bins}
\usage{
bin(x, y, bins = 400, wt)
}
\arguments{
\item{x}{numeric vector or matrix of 2 dimensons of x data}

\item{y}{numeric vector of y data corresponding to \code{x}}

\item{bins}{number of bins to split the data into}

\item{wt}{weight of each observation}
}
\value{
returns an S3 object of the class 'bin' containing
\itemize{
   \item {\code{x} a numeric vector of x data after binned}
   \item {\code{y} a numeric vector of y data corresponding to binned \code{x}}
   \item {\code{weight} weight corresponding to binned \code{x}}
}
}
\description{
Splitting 1D/2D data into bins
}
\examples{
n <- 1000
x <- seq(0,10,length.out = n)
x1 <- rnorm(n,0,0.2)
y <- sin(x) + x1
w <- rep(1/n, n)
z <- bin(x,y,bins=400, w)
}
