% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_selection.R
\name{bw_cv}
\alias{bw_cv}
\title{LOOCV bandwidth selection}
\usage{
bw_cv(
  XY,
  method = c("approx", "exact"),
  kernel = "epanechnikov",
  epsilon = 0.05,
  bw,
  N_min = 1
)
}
\arguments{
\item{XY}{a matrix of the estimators with the observed values in the last colum}

\item{method}{estimation method}

\item{kernel}{type of kernel used}

\item{epsilon}{level of error allowed}

\item{bw}{bandwidth for estimation of local linear regression}

\item{N_min}{number of points stored in the leaf of the tree (only works for approx)}
}
\description{
choosing the best bandwidth through LOOCV
}
