% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcv.llr.R
\name{gcv.llr}
\alias{gcv.llr}
\title{Bandwidth selection using generalized cross validation}
\usage{
gcv.llr(
  x,
  y,
  weight,
  kernel = "epanechnikov",
  approx = FALSE,
  epsilon = 0.05,
  N_min = 1,
  bw
)
}
\arguments{
\item{x}{a numeric vector or matrix of x data}

\item{y}{a numeric vector of y data corresponding to \code{x} data}

\item{weight}{a numeric vector of \code{length(x)} for weight of each data point.}

\item{kernel}{kernel type used; supported are 'epanechnikov', "rectangular", "triangular", "quartic", "triweight", "tricube", "cosine", "gauss", "logistic". "sigmoid" and "silverman".}

\item{approx}{boolean flag: if \code{true} kdtree approximation is used.}

\item{epsilon}{margin of error allowed for llr approximation using kdtree. Only used when \code{approx = TRUE}.}

\item{N_min}{minimum number of points stored in the kd-tree. Only used when \code{approx} = TRUE.}

\item{bw}{a numeric vector or matrix of bandwidth considered for selection.}
}
\value{
returns a single or numeric vector of \code{bw} that gives the smallest mean square error.
}
\description{
Bandwidth selection using generalized cross validation
}
\examples{
  
\dontrun{
n <- 1000
x <- seq(0,10,length.out = n)
x1 <- rnorm(n,0,0.2)
y <- sin(x) + x1
w <- rep(1/n, n)
bw <- seq(0.02,0.4, by =0.01)
binned <- bin(x, y, bins=400, w)
## Bandwidth selection of binned data
h_bin <- gcv.llr(binned$x, binned$y, binned$weight, bw = bw)
## Bandwidth selection of exact local linear regression 
h_exact <- gcv.llr(x, y, w, bw = bw)
## Bandwidth selection of approx local linear regression with kdtree
h_kdapprox <- gcv.llr(x, y, w, approx = TRUE, bw = bw) 
}
}
