% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{ll}
\alias{ll}
\title{Local linear regression}
\usage{
ll(
  X,
  Y,
  method = c("approx", "exact"),
  kernel = "epanechnikov",
  epsilon = 0.05,
  bw,
  N_min = 1
)
}
\arguments{
\item{X}{a matrix of the estimators with the observed values in the last column}

\item{Y}{a matrix of the estimators with the observed values in the last column}

\item{method}{estimation method}

\item{kernel}{type of kernel used}

\item{epsilon}{level of error allowed}

\item{bw}{bandwidth for estimation of local linear regression}

\item{N_min}{number of points stored in the leaf of the tree (only works for approx)}
}
\value{
h
}
\description{
Multivariate local linear regression
}
