% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_predict.R
\name{ll_predict}
\alias{ll_predict}
\title{Local linear regression prediction}
\usage{
ll_predict(
  XY,
  X_pred,
  method = c("approx", "exact"),
  kernel = "epanechnikov",
  epsilon = 0.05,
  bw,
  N_min = 1
)
}
\arguments{
\item{XY}{a matrix of the estimators with the observed values in the last column}

\item{X_pred}{a vector of predictions for observed values}

\item{method}{estimation method}

\item{kernel}{type of kernel used}

\item{epsilon}{level of error allowed}

\item{bw}{bandwidth for estimation of local linear regression}

\item{N_min}{number of points stored in the leaf of the tree (only works for approx)}
}
\description{
Predict values through local linear smoothing
}
