% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.llr}
\alias{plot.llr}
\alias{plot.bin}
\title{Plot class 'llr' object}
\usage{
\method{plot}{llr}(x, xorig, yorig, ...)

\method{plot}{bin}(x, ...)
}
\arguments{
\item{x}{a object of class 'bin'}

\item{xorig}{initial numeric vector x}

\item{yorig}{initial numeric vector y}

\item{...}{for consistency}
}
\description{
Plot class 'llr' object

Plot 'bin' object
}
