% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict1dd}
\alias{predict1dd}
\title{Multiplies two doubles}
\usage{
predict1dd(X, Y, X_pred, kcode, h)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{X_pred}{X_pred}

\item{kcode}{kcode}

\item{h}{h}
}
\description{
Multiplies two doubles
}
