% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Set number of threads to use}
\usage{
set_num_threads(threads)
}
\arguments{
\item{threads}{number of threads to use}
}
\value{
number of threads set
}
\description{
Set number of threads to use
}
