% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Splitting the data into bins}
\usage{
bin(x, y, bins = 400, weight)
}
\arguments{
\item{x}{numeric vector or matrix of 2 dimensons of x data}

\item{y}{numeric vector of y data with length corresponding to \code{x}}

\item{bins}{number of bins to split the data}

\item{weight}{numeric vector corresponding to weight of each observation}
}
\value{
returns an S3 object of the class 'bin' containing
\describe{
   \item{x}{a numeric vector of x data of size \code{bins}}
   \item{y}{a numeric vector of y data corresponding to 'bin' class \code{x}}
   \item{weight}{weight corresponding to 'bin' class \code{x}}
}
}
\description{
Splitting the data into bins
}
\examples{
n <- 1000
x <- seq(0,10,length.out = n)
x1 <- rnorm(n,0,0.2)
y <- sin(x) + x1
w <- rep(1/n, n)
binned <- bin(x,y,bins=400, w)
}
